Implement web module restructuring
- Create web_ui module containing assets.rs for frontend code
- Create web_server module containing server.rs, websocket.rs, and tests.rs
- Update main.rs imports to use new module structure
- Fix test imports to reference correct asset paths
- Remove old web directory after successful migration
- All tests pass and code follows project standards
