+++ src/web_server/mod.rs
@@ -1,4 +1,3 @@
-pub mod assets;
 pub mod server;
 pub mod websocket;
