+++ src/web_server/tests.rs
@@ -1,6 +1,7 @@
 use super::server::WebServer;
 use crate::agent::Agent;
 use crate::ruler::config::{MonitorConfig, WebUIConfig};
+use crate::web_ui::assets;
 use std::sync::Arc;
 
 #[tokio::test]
@@ -37,21 +38,21 @@ fn test_monitor_config_with_web_ui() {
 
 #[test]
 fn test_assets_html_not_empty() {
-    assert!(!crate::web::assets::INDEX_HTML.is_empty());
-    assert!(crate::web::assets::INDEX_HTML.contains("Rule Agents Terminal"));
-    assert!(crate::web::assets::INDEX_HTML.contains("terminal-client.js"));
+    assert!(!assets::INDEX_HTML.is_empty());
+    assert!(assets::INDEX_HTML.contains("Rule Agents Terminal"));
+    assert!(assets::INDEX_HTML.contains("terminal-client.js"));
 }
 
 #[test]
 fn test_assets_css_not_empty() {
-    assert!(!crate::web::assets::MAIN_CSS.is_empty());
-    assert!(crate::web::assets::MAIN_CSS.contains("body"));
-    assert!(crate::web::assets::MAIN_CSS.contains("#282a36"));
+    assert!(!assets::MAIN_CSS.is_empty());
+    assert!(assets::MAIN_CSS.contains("body"));
+    assert!(assets::MAIN_CSS.contains("#282a36"));
 }
 
 #[test]
 fn test_assets_js_not_empty() {
-    assert!(!crate::web::assets::TERMINAL_CLIENT_JS.is_empty());
-    assert!(crate::web::assets::TERMINAL_CLIENT_JS.contains("TerminalClient"));
-    assert!(crate::web::assets::TERMINAL_CLIENT_JS.contains("WebSocket"));
+    assert!(!assets::TERMINAL_CLIENT_JS.is_empty());
+    assert!(assets::TERMINAL_CLIENT_JS.contains("TerminalClient"));
+    assert!(assets::TERMINAL_CLIENT_JS.contains("WebSocket"));
 }
