+++ src/web_ui/assets.rs
